Time: 166 ms (79.49%), Space: 107.6 MB (14.33%) - LeetHub

+++ 2348-number-of-zero-filled-subarrays/2348-number-of-zero-filled-subarrays.cpp
@@ -7,7 +7,7 @@ public:
         for(int i=0;i<n;i++){
             if(a[i]==0){
                 st++;
-                if(i==n-1){ans += (long long)(st*(st+1))/2;}
+                if(i==n-1){ans += (long long)(st*(long long)(st+1))/2;}
             }
             else if(st>0){
                 ans += (long long)(st*(st+1))/2;
